feat: user dto

+++ src/users/users.controller.ts
@@ -1,46 +1,56 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, Post, Put } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpCode,
+  HttpException,
+  HttpStatus,
+  Param,
+  ParseUUIDPipe,
+  Post,
+  Put,
+} from '@nestjs/common';
+
 import { UsersService } from './users.service';
+import { CreateUserDto, UpdateUserDto } from './user.dto';
 
 @Controller('users')
 export class UsersController {
   constructor(private readonly userService: UsersService) {}
 
   @Get()
-  @HttpCode(200)
   getUsers() {
     return this.userService.getUsers();
   }
 
   @Get(':id')
-  @HttpCode(200)
-  getUser(@Param('id') id: string) {
-    return this.userService.getUser(id);
+  getUser(@Param('id', new ParseUUIDPipe()) id: string) {
+    const user = this.userService.getUser(id);
+    if (user) return user;
+
+    throw new HttpException('User not found', HttpStatus.NOT_FOUND);
   }
 
   @Post()
-  @HttpCode(201)
-  createUser(
-    @Body('username') username: string,
-    @Body('age') age: number,
-    @Body('hobbies') hobbies: string[],
-  ) {
-    return this.userService.createUser({ username, age, hobbies });
+  createUser(@Body() createUserDto: CreateUserDto) {
+    return this.userService.createUser(createUserDto);
   }
 
   @Put(':id')
-  @HttpCode(200)
-  updateUser(
-    @Param('id') id: string,
-    @Body('username') username?: string,
-    @Body('age') age?: number,
-    @Body('hobbies') hobbies?: string[],
-  ) {
-    return this.userService.updateUser(id, { username, age, hobbies });
+  updateUser(@Param('id', new ParseUUIDPipe()) id: string, @Body() updateUserDto: UpdateUserDto) {
+    const user = this.userService.getUser(id);
+    if (user) return this.userService.updateUser(id, updateUserDto);
+
+    throw new HttpException('User not found', HttpStatus.NOT_FOUND);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  deleteUser(@Param('id') id: string) {
-    return this.userService.deleteUser(id);
+  deleteUser(@Param('id', new ParseUUIDPipe()) id: string) {
+    const user = this.userService.getUser(id);
+    if (user) return this.userService.deleteUser(id);
+
+    throw new HttpException('User not found', HttpStatus.NOT_FOUND);
   }
 }
